requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
psycopg2-binary
python-dotenv
httpx
groq
python-dotenv
reportlab
PyPDF2
python-docx
